+++ js/main.js
@@ -1,5 +1,6 @@
-import './show-mini-photos.js';
+import './render-mini-photos.js';
 import './show-modal-photo.js';
 import './scale-control.js';
 import './effect.js';
 import './form-validate.js';
+import './server.js';
